+++ controllers/contacts.js
@@ -81,14 +81,9 @@ async function addContact(req, res, next) {
 
 async function updateContact(req, res, next) {
     const { contactId } = req.params; 
-    const contact = {
-        name: req.body.name,
-        email: req.body.email,
-        phone: req.body.phone,
-        favorite: req.body.favorite
-    }
+   
     try {
-        const result = Contact.findByIdAndUpdate(contactId, contact)
+         const result = Contact.findByIdAndUpdate(contactId, req.body)
         res.json({
             status: "success",
             code: 201,
