+++ models/contacts.js
@@ -1,24 +1,6 @@
-// const fs = require('fs/promises')
 
-const listContacts = async () => {}
-
-const getContactById = async (contactId) => {}
-
-const removeContact = async (contactId) => {}
-
-const addContact = async (body) => {}
-
-const updateContact = async (contactId, body) => {}
-
-module.exports = {
-  listContacts,
-  getContactById,
-  removeContact,
-  addContact,
-  updateContact,
-}
 const mongoose = require('mongoose');
-const Joi = require("joi");
+
 
 const cocntactShema = new mongoose.Schema({
     name: {
@@ -37,12 +19,6 @@ const cocntactShema = new mongoose.Schema({
     },
 });
 
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string(),
-  phone: Joi.string(),
-  favorite: Joi.boolean(),
-});
+
 
 module.exports = mongoose.model("Contact", cocntactShema)
-module.exports = {addSchema}
