+++ validation/contactValidationSchemas.js
@@ -1,0 +1,10 @@
+const Joi = require("joi");
+
+const addSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string(),
+  phone: Joi.string(),
+  favorite: Joi.boolean(),
+});
+
+module.exports = {addSchema}
